Fitness center management
It is a python based program with the integration of MySQL that helps in keeping track of all the customers and coaches in FeeltheBurns Fitness centre. It can add new customers, keep track of salaries and fees of the coaches and customers respectively. I does so by taking the email ID and password and assigning a unique ID to all the members of the Centre